% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCriteria.R
\name{loadCriteria}
\alias{loadCriteria}
\title{Load Criteria - the function loads the criteria table
based on the state and tribe name and link to the public GitHub space}
\usage{
loadCriteria(state_tribe, ref)
}
\arguments{
\item{state_tribe}{the state or tribe name}

\item{ref}{the reference table showing the state and tribe name
and their file name}
}
\value{
the criteria table as a data frame
}
\description{
Load Criteria - the function loads the criteria table
based on the state and tribe name and link to the public GitHub space
}
