% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{runAllValidations}
\alias{runAllValidations}
\title{Runs All Validation Functions for a filled out criteria table}
\usage{
runAllValidations(data, validators = NULL, stop_on_error = FALSE)
}
\arguments{
\item{data}{a R data frame with a TADA-compatible criteria table filled out.}

\item{validators}{a character list consisting of the TADACommunityHub
validation functions to run. Default is null which will run all validation
functions.}
}
\value{
A list returning if all seasons are current valid
domain values or not. If not, identify which are not valid.
}
\description{
Runs all validation functions by default. Users can choose to select
which validation functions to run with the validators argument input.
}
\examples{
runAllValidations(UTAHDWQ)

}
