% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UTAHDWQ}
\alias{UTAHDWQ}
\title{UTAHDWQ Criteria and Methodology TADA Table}
\format{
A data frame with 324 rows and 31 columns
}
\description{
A filled out criteria table for UTAHDWQ:
}
\keyword{dataframe}
