% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{validateATTAINSParam}
\alias{validateATTAINSParam}
\title{Load User Data - Validate ATTAINS.ParameterName}
\usage{
validateATTAINSParam(data)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A list returning if all ATTAINS.ParameterName are current valid
domain values or not. If not, identify which are not valid.
}
\description{
Loads a data frame provided by the user.
}
\examples{
data("UTAHDWQ")
validateATTAINSParam(UTAHDWQ)

}
