% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{validateDurationUnits}
\alias{validateDurationUnits}
\title{Load User Data - Validate Duration Units}
\usage{
validateDurationUnits(data)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A list returning if all Duration units are current valid
domain values or not. If not, identify which are not valid.
}
\description{
Loads a data frame provided by the user.
}
\examples{
data("UTAHDWQ")
validateDurationUnits(UTAHDWQ)

}
