% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{validateSeason}
\alias{validateSeason}
\title{Load User Data - Validate Season}
\usage{
validateSeason(data)

validateSeason(data)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A list returning if all seasons are current valid
domain values or not. If not, identify which are not valid.

A list returning if all seasons are current valid
domain values or not. If not, identify which are not valid.
}
\description{
Loads a data frame provided by the user.

Loads a data frame provided by the user.
}
\examples{
validateSeason(UTAHDWQ)

validateSeason(UTAHDWQ)

}
