% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{exportErrors}
\alias{exportErrors}
\title{Export data with errors}
\usage{
exportErrors(data, folder_path = NULL)
}
\arguments{
\item{data}{a single or a list of multiple data frame that is an output from
a TADACommunityHub R validate function.}
}
\value{
An excel spreadsheet that shows the invalid column values from the
user supplied criteria table(s). Users can choose from a drop down list of
allowable valid values for that column name.
}
\description{
Loads the list of unique errors in a column and exports it to df
}
\examples{
review2 <- validateAll(folder_path = "inst/extdata/", validateColumn = validateATTAINS.UseName)
err <- exportErrors(review2)

}
