% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{exportErrors}
\alias{exportErrors}
\title{Export data with errors}
\usage{
exportErrors(data, df_return = TRUE)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A list of list of what column name contains the current valid
domain values or not. If not, identify which are not valid.
}
\description{
Loads the list of unique errors in a column and exports it to df
}
\examples{
data("UTAHDWQ")
review <- validateAll(folder_path = "inst/extdata/", validateColumn = validateWQXChar)

}
