% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{validateAll}
\alias{validateAll}
\title{Validate all data .xlsx in a Folder Path}
\usage{
validateAll(folder_path = NULL, validateColumn)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A list of list of what column name contains the current valid
domain values or not. If not, identify which are not valid.
}
\description{
Loads a data frame provided by the user.
}
