% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadNewData.R
\name{validateATTAINSParam}
\alias{validateATTAINSParam}
\title{Load User Data - Validate ATTAINS.ParameterName}
\usage{
validateATTAINSParam(data)
}
\arguments{
\item{data}{a R data frame. Future dev will allow other data file types.}
}
\value{
A data frame containing the loaded data.
}
\description{
Loads a data frame provided by the user.
}
